Make ResultPartitionWriter record-oriented
Currently, the ResultPartitionWriter is buffer-oriented, that is, RecordWriter can only add Buffers of different channels to it, which is not flexible enough for new implementations like sort-based partitioning. This patch aims to make ResultPartitionWriter record-oriented by adding new record-oriented interfaces to and removing the old buffer-oriented from ResultPartitionWriter. After this change, the sort-based result partition can be implemented.