Introduce BufferWritingResultPartition and migrate logics related to ResultSubpartition from ResultPartition to it
The BufferWritingResultPartition writes buffers directly to ResultSubpartitions. This is in contrast to implementations where records are written to a joint structure, from which the data is drawn after the write phase is finished, for example the sort-based partitioning. In addition, the logics related to ResultSubpartition are moved from ResultPartition to BufferWritingResultPartition for these logics can not be shared by the sort-based result partition.