[FLINK-19312][network] Introduce BufferWritingResultPartition which wraps the ResultSubpartition related logic
In the current abstraction, buffers are written to and read from ResultSubpartitions which is a hash-style data writing and reading implementation. This is in contrast to implementations where records are appended to a joint structure, from which the data is drawn after the write phase is finished, for example the sort-based partitioning which clusters data belonging to different channels by sorting channel index. In the future, sort-merge based ResultPartitionWriter will be implemented which can not share the current hash-style ResultSubpartition related logics. So this patch migrates these logics from ResultPartition to the new BufferWritingResultPartition, after which the ResultPartition is free of ResultSubpartition and can be reused by the future sort-merge based ResultPartitionWriter implementation.