[FLINK-19320][task] Remove RecordWriter#clearBuffers
Previously, RecordWriter#clearBuffers was used to recycle the partially filled buffer in the serializer. However, the serializer does not contain any network buffer any more so currently, the method is only used to finish the current BufferBuilder and only some tests and BatchTask use it. Actually, these usage should be replaced by RecordWriter#close which dose the same thing. So this patch removes RecordWriter#clearBuffers and the corresponding test cases.