Avoid affects the multiple inputs & deadlock detection
This fix tries to fix the issue that if we mark dpp as
the input of the table scan source, it will be detected
as circles in Deadlock detection processor and insert
another Exchange between the fact source scan and the
multiple input operators, which forbid the fact source
get chained.

The current fix first remove the inputs between dpp sink and
the source, then at last it tries to see if the fact source
seems to be able to be chained. If so, it skip adding the dependency
edge, otherwise the dependency edge is added.

The HiveDialectITCase is the case wiout multiple input nodes, which
still needs the dependency edge. The q1.sql is an example of the dependency
edge could be omitted.

It seems we could not easily remove the inputs in the physical layer for
two reasons:
1. Without the input, the dynamic filter collector will be duplicated,
one for the fact source, one for the normal dim->calc->dpp collector.
2. Without the input, when checking subgraph reuse, it could not
consider the condition (which is stored in the calc node of dim -> calc
-> dpp), then if there are two fact scans with the same dim with
different conditions, the dim/dpp might be reused wrongly (See q4.sql or
HiveDialectITCase#testDpp2).