Avoid duplication dynamic partition collectors
First to translate the dpp collector exec node, we should
use translate instead of translateInternal to avoid duplication
the transformations.

With translate() previously the code works correctly in a wierd way
if the same dim are reused with multiple fact scan:
1. When translating the first scan, the dpp collector transformation will
be created with the given uuid.
2. Notes that the dpp collector exec node passed the dynamicPartitionIds
list directly to the operator factory, namely they are sharing the same
object.
3. When translating the second scan, the dpp collector transformation
is reused, and by registering the second uuid to the list in the dpp
collector exec node, we are in fact also passing it to the dpp collector
operator factory.

Althought it works, it should be very fragile. We might explicitly add
the uuid to the operator factory directly.